fixed process image and retrieve results functions
- Split original lambda into two functions
